requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
pydantic==2.7.0
psycopg2-binary==2.9.9
redis==5.0.4
celery==5.4.0
requests==2.32.0
beautifulsoup4==4.12.3
groq==0.13.0
python-dotenv==1.0.1
